requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==0.28.0
openai-whisper==20250625
pillow==11.3.0
propcache==0.3.2
PyAutoGUI==0.9.54
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.1
python-telegram-bot==22.2
pytweening==1.2.0
pywhatkit==5.4
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wikipedia==1.4.0
yarl==1.20.1
